+++ plugins/clickhouse/src/components/panel/LogsDocuments.tsx
@@ -1,4 +1,12 @@
-import { TableComposable, TableVariant, Th, Thead, Tr } from '@patternfly/react-table';
+import {
+  IExtraColumnData,
+  SortByDirection,
+  TableComposable,
+  TableVariant,
+  Th,
+  Thead,
+  Tr,
+} from '@patternfly/react-table';
 import React from 'react';
 
 import { IDocument } from '../../utils/interfaces';
@@ -7,16 +15,25 @@ import LogsDocument from './LogsDocument';
 interface ILogsDocumentsProps {
   documents?: IDocument[];
   fields?: string[];
+  order?: string;
+  orderBy?: string;
   addFilter?: (filter: string) => void;
+  changeOrder?: (order: string, orderBy: string) => void;
   selectField?: (field: string) => void;
 }
 
 const LogsDocuments: React.FunctionComponent<ILogsDocumentsProps> = ({
   documents,
   fields,
+  order,
+  orderBy,
   addFilter,
+  changeOrder,
   selectField,
 }: ILogsDocumentsProps) => {
+  const activeSortIndex = fields && orderBy ? fields?.indexOf(orderBy) : -1;
+  const activeSortDirection = order === 'descending' ? 'desc' : 'asc';
+
   return (
     <TableComposable aria-label="Logs" variant={TableVariant.compact} borders={false}>
       <Thead>
@@ -24,7 +41,27 @@ const LogsDocuments: React.FunctionComponent<ILogsDocumentsProps> = ({
           <Th />
           <Th>Time</Th>
           {fields && fields.length > 0 ? (
-            fields.map((selectedField, index) => <Th key={index}>{selectedField}</Th>)
+            fields.map((field, index) => (
+              <Th
+                key={index}
+                sort={{
+                  columnIndex: index,
+                  onSort: (
+                    event: React.MouseEvent,
+                    columnIndex: number,
+                    sortByDirection: SortByDirection,
+                    extraData: IExtraColumnData,
+                  ): void => {
+                    if (changeOrder) {
+                      changeOrder(sortByDirection === 'desc' ? 'descending' : 'ascending', field);
+                    }
+                  },
+                  sortBy: { direction: activeSortDirection, index: activeSortIndex },
+                }}
+              >
+                {field}
+              </Th>
+            ))
           ) : (
             <Th>Log</Th>
           )}